+++ Solutions/16priyesh/1_16priyesh.cpp
