+++ Solutions/AK-007/1_AK-007.py
