+++ Solutions/Jigar3/1 _Jigar3.c
