+++ Solutions/Tachyon01/1_Tachyon01.c
