+++ Solutions/muskan6769/1_muskan6769.cpp
