+++ Solutions/nikki1016/1_nikki1016.cpp
