+++ Solutions/poor-kid/1_poor-kid.cpp
