+++ Solutions/pratham29/1_pratham29.cpp
