+++ Solutions/shyam49/1_shyam49.cpp
